refactor: 유효하지 않은 FCM 토큰 처리 로직 분리 및 최적화 #326
