+++ src/main/java/com/hrr/backend/domain/fcm/service/FcmTokenDeactivationService.java
@@ -1,0 +1,11 @@
+package com.hrr.backend.domain.fcm.service;
+
+import com.google.firebase.messaging.BatchResponse;
+import java.util.List;
+
+public interface FcmTokenDeactivationService {
+    /**
+     * FCM 발송 결과에서 실패한 토큰들이 유효하지 않은 경우 비활성화 처리
+     */
+    void handleFailedTokens(List<String> tokens, BatchResponse response);
+}
